Use fake sendmail script for TravisCI builds
This avoids 'sh: 1: /usr/sbin/sendmail: not found' error message
displayed after PHPUnit suite execution, as sendmail is not installed on
Travis.

Note: This is not a "real" error and does not actually cause the build
to fail, but it could lead someone reviewing the build's log to think
that something went wrong, so it should be fixed to avoid confusion.

The fakesendmail.sh script was copied from PHPMailer's test suite [1].
It will generate a .eml file in /tmp/fakemail directory, each time
the test suite sends an email.

Fixes #32828

[1]: https://github.com/PHPMailer/PHPMailer/blob/v6.8.0/test/fakesendmail.sh