Fix testProjectUpdateVersion()
The assertion to test the version update (PATCH request) was incorrectly
setting the timestamp in the updated version data as a Unix timestamp
instead of a date string to be parsed by strtotime().

Depending on the current time, the timestamp is sometimes converted by
the strtotime() call in VersionUpdateCommand to a date that is not
supported by the database which causes the SQL UPDATE statement in
version_update() to fail.

Fixes #33374