Use enabled/disabled for category status
Since we currently only use a binary state (enabled/disabled) for the
status, we do not need to make things more complex than they have to be.

This simplifies @cas's original contribution by:
- using a checkbox for input instead of radio buttons
- display a check icon instead of relying on language strings in lists
- standardize on enabled/disabled terminology

This also fixes the "inverted logic" (status == 0 was previously treated
as enabled, which is the reverse of one would expect.

Fixes #31017