Keep current disabled category when editing issue
When an issue assigned to a disabled category is edited, the category
should be kept as-is unless the user specifically updated it.

Update print_category_option_list() to add the current category, but
label it as disabled so the user knows they should change it, and set
the disabled attribute on the option so it can't be re-selected once
changed.

Fixes #31017