Use absolute path to schema.php in html API
print_admin_menu_bar() needs to include schema.php, to determine whether
we have an upgrade pending or not.

This include_once call did not specify the path to the file, resulting
in a failure to include it on IIS with FastCGI, as the current working
directory was not set to that of the currently executing script, but to
where the PHP executable was stored. This caused $g_upgrade to be null,
leading to an error when calling count() on PHP 8.

We now use require_once instead so we get a hard failure if schema.php
could not be included, and we specify the full path to the file to avoid
the problem on IIS/FastCGI.

Fixes #33906