Fix workflow labels with space displayed as HTML
The initial attempt to fix this by generating an HTML-like label failed,
because GraphViz's "HTML-like" labels only allow a limited set of tags
and the `<span class="nowrap">` inserted by string_nobreak() is not one
of them.

So the only way to fix this is to remove that function call, which
should not cause any problems, because Graphviz will not wrap text
unless a newline `\n` is inserted in the label.

Fixes #34607