Use helper_mantis_url(), not lang_get_defaulted()
The incorrect usage is a regression introduced many years ago by commit
b8e0279f7c1026bb27aaa5a1f3f09e5b9d8b08fc.

It never caused any noticeable harm, because with lang_get_defaulted()
the param is returned as-is when not found, and since it was used as URL
it would be relative to current directory.

Fixes #34682