Use a stacked y axis for Issue Trend chart
The modified visualization introduced in commit  c54672ca3d2058dfb0b2dfa4843ab0196335227d
was confusing, due to the `Still open` dataset being the delta of
`Reported` - `Resolved`.

We now display the areas under the `resolved` and `still open` datasets
as stacked data. Since the sum of the 2 is equal to the `reported` set,
its line overlaps and we avoid confusing display.

Fixes #34847