New config_is_set_in_database() API function
Same as existing config_is_set(), except that it returns false if there
is not override for the config in the database (i.e. it does not check
the default value from config files.

This is is needed so that PHPUnit tests are able to change config values
and then restore the database to its previous state afterwards.

Fixes #35211