New top-level abstract base class MantisTestCase
To be used as parent class for RestBase, SoapBase and MantisCoreBase to
avoid duplicating common helper methods.

This was originally meant to move the getTestCaseReference() method, but
since it was finally removed after a better solution was found with
PHPUnit native toString() method [1], the class is currently an empty
shell.

Committing the change anyway, as it may prove useful in the future.

[1]: PR https://github.com/mantisbt/mantisbt/pull/20742074