Install: relax severity of failed $g_path test
In some cases (e.g. redirections, authenticated pages), url_get() fails
to retrieve the MantisBT home page. This blocks the installer, because
the check is marked as "hard fail".

This commit relaxes the check to a simple warning, making it the Admin's
responsibility fix the system if it ends up to be non-functional due
to an invalid $g_path.

Fixes #34783